ch3_github_issues/pyproject.toml
[project]
name = "ch3_github_issues"
version = "0.1.0"
description = "ch3-github-issues using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.119.0,<1.0.0",
    "PyGithub>=2.1.1"
]

[project.scripts]
ch3_github_issues = "ch3_github_issues.main:run"
run_crew = "ch3_github_issues.main:run"
train = "ch3_github_issues.main:train"
replay = "ch3_github_issues.main:replay"
test = "ch3_github_issues.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
